Fix for rare case when a single-push p2sh input leads to crash reporter
Fixes issue #2958.

Transactions.parse_scriptSig sometimes mismatches p2sh inputs that have
a single push as "p2pk", and the rest of the app doesn't know how to
deal with this situation, mainly due to uncaught exception from
xpubkey_to_address. The fix is to catch this exception and handle the
situation correctly.